+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- In admin "show-details" sidebar, make the linked resources count clickable
+
 ## [0.5.1] - 2023-11-23
 
 - Fixed AJAX calls when Taxonomy is the only enabled module doing it (enabled