+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modify admin and site advanced search forms to be able to search resources
   that links to a specific taxonomy term or its descendants.
 - On taxonomy term pages, show resources linked to term descendants too
+- Admin taxonomy and taxonomy term pages now show links to the corresponding
+  sites pages
 
 ## [0.5.1] - 2023-11-23