+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.6.0] - 2023-12-05
 
 - In admin "show-details" sidebar, make the linked resources count clickable
 - Show children count on term show page and in show-details sidebar
@@ -72,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
+[0.6.0]: https://github.com/biblibre/omeka-s-module-Taxonomy/releases/tag/v0.6.0
 [0.5.1]: https://github.com/biblibre/omeka-s-module-Taxonomy/releases/tag/v0.5.1
 [0.5.0]: https://github.com/biblibre/omeka-s-module-Taxonomy/releases/tag/v0.5.0
 [0.4.0]: https://github.com/biblibre/omeka-s-module-Taxonomy/releases/tag/v0.4.0